wing_tuning: lift form state into Pinia store + fix angle_pitch_offset tooltip
Move the 26-field `fields` reactive (plus FIELD_DEFS / defaultFields)
out of WingTuningTab.vue into useWingTuningStore, making the store the
single source of truth for form state — mirrors the state-holder
pattern in betaflight/blackbox-log-viewer#913 (src/stores/log.js).
MSP I/O stays in the tab; storeOriginals() now snapshots the store's
own `fields` instead of FC.WING_TUNING.

Also fixes an inverted tooltip: angle_pitch_offset positive pitches
the nose down, negative pitches it up (per firmware pid.h:327).
